--- internal/source/mock_source.go
@@ -1,11 +1,0 @@
-package source
-
-type MockSource struct {
-	Source chan any
-}
-
-func (receiver *MockSource) ForwardTo(target chan any) {
-	for update := range receiver.Source {
-		target <- update
-	}
-}
